PSP: analog stick works, level loads on hardware -- clock, paths, WAV alignment
- The stick never fed the game: the SDL input driver scales axes by the
  frame delta, kept as a FLOAT of absolute seconds, and the PSP counter is
  microseconds since the year 1 (~6e10 s), so every delta rounded to zero.
  appSeconds() is now relative to startup on PSP and the input timestamp is
  a DOUBLE. Stick speed 0.5, dead zone 0.2, vertical look inverted.
- New Game said "missing file" on hardware: the menu URL is
  "..\maps\Vortex2.unr" and sceIo takes neither backslashes nor "..".
  appFopen/appFSize now normalise paths (slashes, cwd-anchored, ".."
  collapsed). PPSSPP quietly accepted the raw path.
- The level change then died with an address error in ReadWaveInfo: the
  'smpl' chunk's cSampleLoops is read through a struct overlaid on file
  data. The WAV chunk structs are packed on PSP so every field access is
  byte-safe.
- First stick events are logged (PSPJOY) for tuning.

+++ Source/NSDLDrv/Inc/NSDLDrv.h
@@ -75,7 +75,7 @@ private:
 	SDL_Rect DisplaySize;
 	SWORD JoyAxis[SDL_CONTROLLER_AXIS_MAX];
 	UBOOL QuitRequested;
-	FLOAT InputUpdateTime;
+	DOUBLE InputUpdateTime;   // DOUBLE: a FLOAT of absolute seconds cannot resolve a frame
 
 	// Info saved during captures and fullscreen sessions.
 	INT SavedX, SavedY;
